change seven_point_rating type to integer

+++ db/migrate/20230718021133_add_seven_point_rating_to_surfline_v2s.rb
@@ -1,6 +1,6 @@
 class AddSevenPointRatingToSurflineV2s < ActiveRecord::Migration[7.0]
   def change
-    add_column :surfline_v2_lolas, :seven_point_rating, :decimal
-    add_column :surfline_v2_lotus, :seven_point_rating, :decimal
+    add_column :surfline_v2_lolas, :seven_point_rating, :integer
+    add_column :surfline_v2_lotus, :seven_point_rating, :integer
   end
 end
